Refactored some of the logic to views
