+++ Assets/ImportExport/Views/IDExportView.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1e5c21ebaf8383b4ea044503b6564676
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
