+++ Assets/ImportExport/Views/PrefabView.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: eb605ef8a909df840af5cd37f71714ed
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
