+++ Assets/ImportExport/Views/SceneView.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ad7d99f0c4c180b439025471175c28a2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
